Moved the client and cleaned up the imports to be much more simple and explicit - only shortcuts for options and client classes
