+++ devcycle_python_sdk/__init__.py
@@ -1,26 +1,4 @@
-# coding: utf-8
+# Simplify imports for the SDK entry point objects
 
-# flake8: noqa
-
-"""
-    DevCycle Bucketing API
-
-    Documents the DevCycle Bucketing API which provides and API interface to User Bucketing and for generated SDKs.  # noqa: E501
-
-    OpenAPI spec version: 1.0.0
-
-    Generated by: https://github.com/swagger-api/swagger-codegen.git
-"""
-
-from __future__ import absolute_import
-
-# import apis into sdk package
-from devcycle_python_sdk.api.dvc_cloud_client import DVCCloudClient
 from devcycle_python_sdk.dvc_options import DVCCloudOptions
-from devcycle_python_sdk.api.bucketing_client import BucketingAPIClient
-# import models into sdk package
-from devcycle_python_sdk.models.error_response import ErrorResponse
-from devcycle_python_sdk.models.event import Event
-from devcycle_python_sdk.models.feature import Feature
-from devcycle_python_sdk.models.user_data import UserData
-from devcycle_python_sdk.models.variable import Variable, TypeEnum
+from devcycle_python_sdk.dvc_cloud_client import DVCCloudClient
